fix(avalonia): address #792 review — correct B16-B31 sync wording + drop redundant cast
Copilot review on PR #792 flagged 4 minor doc/comment-accuracy items (no
behavior change). The 16 B16..B31 NumericUpDowns are x:Name-only (ExtB16Box…)
with NO AXAML {Binding} — they are synced by code-behind UpdateUI() (push
VM→controls) / WriteButton_Click() (pull controls→VM), the same mechanism as
the editor's other fields. Fixes:

1. README: drop the inaccurate "two-way bound to Ext0..Ext15" phrasing; describe
   the real code-behind UpdateUI/WriteButton_Click sync (in-place, undoable).
2. SkillConfigFE8NSkillView.axaml: reword the grid comment — the inputs are
   x:Name-only, values pushed in UpdateUI() and pulled in WriteButton_Click(),
   no AXAML binding (consistent with the other fields).
3. SkillConfigFE8NSkillViewModel.LoadEntry(): reword the comment — assign
   through the props for consistency with the other fields and correct IsLoading
   dirty-tracking suppression, NOT because of control bindings.
4. SkillConfigFE8NSkillViewModel write loop: remove the redundant (byte) cast in
   rom.write_u8(addr + 16u + i, _ext[i]) — write_u8 takes uint and truncates;
   matches the rest of the method passing uint directly.

No test changes (behavior identical). Verification:
- dotnet build FEBuilderGBA.Avalonia -c Release: 0 errors
- dotnet test FEBuilderGBA.Avalonia.Tests --filter "FullyQualifiedName~SkillConfigFE8N": 97 passed, 0 failed
- pwsh scripts/validate-automation-ids.ps1: exit 0

+++ FEBuilderGBA.Avalonia/ViewModels/SkillConfigFE8NSkillViewModel.cs
@@ -354,8 +354,10 @@ namespace FEBuilderGBA.Avalonia.ViewModels
             CondItem4 = rom.u8(addr + 15);
 
             // Ext-tab bytes B16..B31 (#790 — now editable via Ext0..Ext15).
-            // Assign through the properties so PropertyChanged fires for any
-            // bound NumericUpDowns (IsLoading suppresses dirty marking).
+            // Assign through the properties (not the backing array) for
+            // consistency with the other fields above, so IsLoading dirty-
+            // tracking suppression applies uniformly. The View's UpdateUI()
+            // pushes these into the NumericUpDowns — there is no AXAML binding.
             Ext0 = rom.u8(addr + 16u);
             Ext1 = rom.u8(addr + 17u);
             Ext2 = rom.u8(addr + 18u);
@@ -420,7 +422,7 @@ namespace FEBuilderGBA.Avalonia.ViewModels
             // the +31 bound is already covered by the RowStride-1 guard above.
             for (uint i = 0; i < 16; i++)
             {
-                rom.write_u8(addr + 16u + i, (byte)_ext[i]);
+                rom.write_u8(addr + 16u + i, _ext[i]);
             }
         }
